fix 64 bit warnings, convert to Xcode 5.1, fix unit tests memory leaks

+++ DTCollectionViewManager/DTCollectionViewController.m
@@ -29,7 +29,7 @@
 
 @interface DTCollectionViewController ()
 <DTCollectionFactoryDelegate,UICollectionViewDelegateFlowLayout>
-@property (nonatomic, assign) int currentSearchScope;
+@property (nonatomic, assign) NSInteger currentSearchScope;
 @property (nonatomic, copy) NSString * currentSearchString;
 @property (nonatomic, retain) DTCollectionViewFactory * factory;
 @end
